save account behavior modified
When a request does not provide a UID, we treat as a new item. If it
provides a UID then we should check if the account really exist, so we
can update, or it does not and we return nil and an error.

We also take care of the values that can not be updated: uid, created so
we always have the rigth value.

Also added logxi logger.